Reorganized the server methods to return a string with all the values returned from the queries.
Upgraded the server classes.
The server is now essentially the middle man between the UI code and the DB.
Updated the BookModel to reflect the database entity.
Updated the abstract model to reflect common fields for media objects.
Updated the view with the new values.
Added a lib dir and put the sqlite jar in it
Added the sql jar to the classpath relative to the lib folder
Updated the client to be a object class instead of a main class
Updated the column constants to reflect the new columns of the c3db books table
Added the database file
Added the csv file i used to populate the books table.TABLE